Infer return types for arraynd arithmetic operations
Previously it was assumed the return
type was the same as the operand type,
but this was incorrect for T = quantity<...>.

For example, multiplying two multidimensional
arrays of distances should produce a
multidimensional array of areas (distance
squared).

It was necessary to remove the
const qualifiers on quantity operator
return types to prevent the arraynd
operations from inferring const
array elements.